pyproject.toml
[project]
name = "extracao-infos-pdf-darf-flask-web-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "msal>=1.34.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.1",
    "rapidocr-onnxruntime>=1.3.0",
    "werkzeug>=3.1.3",
]
